+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Resizing the window (including rotating) will re-calculate board sizing.
 * When zoomed in with Tenuki's own zoom, pinch-zooming out will zoom the board out.
 * Simplify textured stone styling.
+* Add a new default renderer which uses SVG.
 
 # v0.2.0